Separated Questions topic wise and added more solutions
